fix: refactor logging
